+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Added support for struct enum variants
 - Tagged more functions and methods as `#[inline]`
 - Switched to using GitHub Actions as CI provider
 - Bumped minimum supported Rust version to `1.57`